Initial implementation of recovery from checkpoint. (#411)
* tmp

* Support header only in consensus.

* Merge branch 'master' into header_only

* Handle ignore_body.

* nit.

* Remove on_new_block_construction_only.

* Remove construct_pivot_info().

* Merge branch 'master' into fullnode

* Merge branch 'header_only' into fullnode

* Pass tests.

* Add fullnode test.

* Delay clone to tx pool.

* Merge branch 'master' into header_only

* Merge branch 'header_only' into fullnode

* Remove unused code.

* Merge branch 'master' into fullnode

* fmt.

* Fix some comments.

* Add FullNode.

* Bug fix.

* Bug fix again.

+++ client/src/rpc/impls/cfx.rs
@@ -317,7 +317,7 @@ impl RpcImpl {
         Ok(self.consensus.best_block_hash())
     }
 
-    fn get_block_count(&self) -> RpcResult<usize> {
+    fn get_block_count(&self) -> RpcResult<u64> {
         info!("RPC Request: get_block_count()");
         Ok(self.consensus.block_count())
     }
@@ -901,7 +901,7 @@ impl TestRpc for TestRpcImpl {
         self.rpc_impl.get_best_block_hash()
     }
 
-    fn get_block_count(&self) -> RpcResult<usize> {
+    fn get_block_count(&self) -> RpcResult<u64> {
         self.rpc_impl.get_block_count()
     }
